Calculator - Exercises 2 and 3 done

+++ Exercises/Calculator/src/SimpleNode.java
@@ -9,7 +9,7 @@ class SimpleNode implements Node {
     protected Object value;
     protected Calculator parser;
 
-    public int val;
+    public float val;
     public int Op = 0;
 
     public SimpleNode(int i) {
